Add documentation and move GridView related components to Wpf.Ui/Controls/
